multiple analyses and huge Map

+++ src/main/java/fr/inrae/act/bagap/chloe/window/counting/BasicCounting.java
@@ -2,7 +2,7 @@ package fr.inrae.act.bagap.chloe.window.counting;
 
 public class BasicCounting extends Counting implements BasicCountingInterface {
 
-	public BasicCounting(float theoreticalSize) {
+	public BasicCounting(double theoreticalSize) {
 		super(theoreticalSize);
 	}
